pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.2.5",
    "googlemaps>=4.10.0",
    "gradio>=5.32.1",
    "langchain-google-genai>=2.1.2",
    "mcp[cli]>=1.9.2",
    "playwright>=1.52.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ipython>=9.3.0",
]
